saleRep refactorized to Long

+++ salesrep-service/src/main/java/com/ironhack/salesrepservice/Repository/SalesRepRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SalesRepRepository extends JpaRepository<SalesRep, Integer> {
+public interface SalesRepRepository extends JpaRepository<SalesRep, Long> {
